requirements.txt
mlserver==1.3.5
ultralytics==8.2.22
torch==2.2.1
torchvision==0.17.1
opencv-python-headless==4.9.0.80
PyYAML==6.0.1
numpy==1.26.4
Pillow==10.2.0
